==> 4.2 - Code Day/4. Indexes.ipynb <==
# Slice and Dice Pandas DataFrames via Indexes

## 1. Import libraries and dependencies

# Import libraries and dependencies
import pandas as pd
from pathlib import Path

## 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csv_path = Path('../4.2 - Code Day/Resources/people_cleansed.csv')

## 3. Read the CSV into a Pandas DataFrame

# Read in the CSV as a DataFrame
people_csv = pd.read_csv(csv_path)
people_csv.head()

## 4. Index Selection with iloc

### 1. Select First Row with iloc

# Select the first row of the DataFrame
people_csv.iloc[0]

### 2. Select First 10 Rows Using Index Selection with iloc

# Select the first 10 rows of the DataFrame
people_csv.iloc[0:10] 

### 3. Select Last Row Using Index Selection with iloc

# Select the last row of the DataFrame
people_csv.iloc[-1]

### 4. Select Second Column Using Index Selection with iloc

# Select the second column of the DataFrame
people_csv.iloc[:,1].head()

### 5. Select Last Column Using Index Selection with iloc

# Select the last column of the DataFrame, with all rows
people_csv.iloc[:,-1].head()

### 6. Select First Three Columns Using Index Selection with iloc

# Select the first three columns of the DataFrame, with all rows
people_csv.iloc[:, 0:3].head()

### 7. Select Specific Rows and Columns Using Index Selection with iloc

# Select the 1st, 3th, 5th, 7th rows of the 2nd, 4th, and 6th columns.
people_csv.iloc[[0,2,4,6], [1,3,5]]

### 8. Select Ranges of Rows and Columns Using Index Selection with iloc

# Select the first 5 rows of the 3rd, 4th, and 5th columns of the DataFrame
people_csv.iloc[0:5, 2:5] 

### 9. Assign Values to Specific Rows and Columns Using Index Assignment with iLoc

# Modify the 'first_name' column value of the first row
people_csv.iloc[0, people_csv.columns.get_loc('First_Name')] = 'Arya'
people_csv.head()

## 5. Index Selection with `loc`:

### 1. Set Column as Pandas DataFrame Index

# Set the index as the 'First_Name' column
people_csv.set_index(people_csv['First_Name'], inplace=True)
people_csv.head()

### 2. Sort Pandas DataFrame by Index

# Sort the index
people_csv.sort_index(inplace=True)
people_csv

### 3. Select Specific Row with Index Selection using loc

# Select the row with the index 'Andrew'
people_csv.loc['Andrew']

### 4. Select Ranges of Rows with Index Selection using loc

# Slice the data to output the range of rows between 'Andrew' and 'Vonnie' using `loc`.
people_csv.loc['Andrew':'Vonnie'].head()

### 5. Select Rows with Conditionals using loc

# Select the rows where `Gender` is equal to 'Male' using `loc`.
people_csv.loc[people_csv['Gender'] == 'Male'].head()

### 6. Assign Values to Specific Rows and Columns Using Index Assignment with loc

# Change the `First_Name` value of the rows with 'Aaron' as the `First_Name` index to 'Arod'
people_csv.loc['Aaron', 'first_name'] = 'Arod'
people_csv.head()
