==> 4.2 - Code Day/5. Sorting.ipynb <==
# Sorting Pandas DataFrames

## 1. Import Libraries and Dependencies

# Import libraries and dependencies
import pandas as pd
from pathlib import Path

## 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csv_path = Path('../4.2 - Code Day/Resources/people_cleansed.csv')

## 3. Read the CSV into a Pandas DataFrame

# Read in the CSV as a DataFrame
people_df = pd.read_csv(csv_path)
people_df.head()

## 4. Sort DataFrame in Ascending Order

# Sort ascending (default)
people_df.sort_values("First_Name")

## 5. Sort DataFrame in Descending Order

# Sort descending
people_df.sort_values("First_Name", ascending=False)
