==> 4.2 - Code Day/6. GroupBy.ipynb <==
# Grouping Data in Pandas DataFrames

## 1. Import Libraries and Dependencies

# Import necessary libraries and dependencies
import pandas as pd
from pathlib import Path
%matplotlib inline

## 2. Create a Path to the File Using Pathlib

# Use the Pathlib libary to set the path to the CSV
csv_path = Path('../4.2 - Code Day/Resources/people_cleansed.csv')

## 3. Read the CSV into a Pandas DataFrame

# Use the file path to read the CSV into a DataFrame and display a few rows
people_df = pd.read_csv(csv_path)
people_df.head()

## 4. Group DataFrame by `Occupation` and perform `count` aggregation

# Group by `Occupation` and perform count
people_df.groupby('Occupation').count()

## 5. Group DataFrame by `Occupation` and Calculate Average Salary and Age

# Calculate average Salary and Age for each Occupation
people_df.groupby('Occupation').mean()

## 6. Group By `Occupation` and `Gender` Columns, then Calculate Average Salary and Age

# Group by Occupation and Gender columns
people_df.groupby(['Occupation','Gender']).mean()
