==> 4.2 - Code Day/7. Binning.ipynb <==
# Categorizing Records in Pandas DataFrames

## 1. Import Libraries and Dependencies

# import dependencies
import pandas as pd
from pathlib import Path

## 2. Set the File Path to the CSV using Pathlib

# Use the Pathlib libary to set the path to the CSV
csv_path = Path('../4.2 - Code Day/Resources/people_cleansed.csv')

## 3. Import the CSV into Pandas DataFrame

# Use the file path to read the CSV into a DataFrame and display a few rows
people_df = pd.read_csv(csv_path)
people_df.head()

## 4. Create Bins for Total Ranking

# Create the variable bins to define our bounds for when we cut the dataframe.
bins = [0, 30000, 70000, 100000, 200000]

# Create names for the bins
group_names = ["Low", "Moderate", "Above Average", "High"]

## 5. Label Records According to the Corresponding Bin

# Add a column named "Salary Level" to people_df and slice the data into the bins
people_df["Salary Level"] = pd.cut(people_df["Salary"], bins, labels=group_names)

people_df.head()
