==> 4.2 - Code Day/10. Concat.ipynb <==
# Concatenating Pandas DataFrames

## 1. Import Libraries and Dependencies

# Import libraries and dependencies
import pandas as pd
from pathlib import Path

## 2. Create a Path to the File Using Pathlib.

# Import data
msft_csv_path = Path('../4.2 - Code Day/Resources/MSFT.csv')
sp500_csv_path = Path('../4.2 - Code Day/Resources/SP500.csv')

## 3. Read the CSV into a Pandas DataFrame.

# Read the CSVs into DataFrames
msft_df = pd.read_csv(msft_csv_path)
sp500_df = pd.read_csv(sp500_csv_path)

## 4. Visualize MSFT DataFrame

# Display a few rows of the MSFT DataFrame
msft_df.head()

## 5. Visualize SP500 DataFrame

# Display a few rows of the SP500 DataFrame
sp500_df.head()

## 6. Concatenate Data by Columns Using `concat` Function and `inner` Join

# Concatenate the two DataFrames by columns and perform an inner join
column_appended_data = pd.concat([msft_df, sp500_df], axis="columns", join="inner")
column_appended_data

## 7. Concatenate Data by Row Using `concat` Function and `inner` Join

# Concatenate the two DataFrames by rows and perform an inner join
column_appended_data = pd.concat([msft_df, sp500_df], axis="rows", join="inner")
column_appended_data
