==> 4.2 - Code Day/16. Culmination.ipynb <==
# World Happiness

## 1. Import Libraries and Dependencies

# Import dependencies
import pandas as pd

## 2. Set File Path to CSV using Pathlib

# Create reference to CSV file
# Hint - the file path may require some extra steps..
csv_path = "../4.2 - Code Day/Resources/world_happiness.csv"

## 3. Import CSV into Pandas DataFrame

# Import the CSV into a pandas DataFrame
happiness_df = pd.read_csv(csv_path, low_memory=False)
happiness_df

## 4. Set Index and Sort Index in Ascending Order

# Set new index to Country
happiness_df = happiness_df.set_index("Country")

# Sort the new dataframe alphabetically
happiness_df.sort_index(inplace=True)
happiness_df.head(30)

## 5. Select the Happiness Score for Vietnam

# Grab the data contained within the row of a country of your choice and the "Happiness.Score" column
# Print the result as "<country's name> happiness score for 2017 was <happiness score>"
happy_score = happiness_df.loc["Vietnam", "Happiness.Score"]
print("Vietnam's happiness score for 2017 was " + str(happy_score))

## 6. Select the Happiness Rank and Score for Belgium, Belize, Benin, Bhutan, and Bolivia

# Grab all rows of data from "Belgium" to "Bolivia" and the columns from "Happiness.Rank" to "Happiness.Score"
# Hint - in order to view more data you will need to increase the dataframe's head
belgium_to_bolivia = happiness_df.loc[["Belgium", "Belize", "Benin",
                                "Bhutan", "Bolivia"], ["Happiness.Rank", "Happiness.Score"]]
belgium_to_bolivia

## 7. Select the Freedom and Government Trust for Every Country

# Select all rows and only display the columns for "Freedom" and "Trust..Government.Corruption." 
happiness_df.loc[:, ["Freedom", "Trust..Government.Corruption."]].head()

## 8. Select all the Countries Where GDP is Greater than or Equal to 1.6

# Select all rows where the GDP per capita has a value greater than or equal to 1.6
gdp_score = happiness_df.loc[happiness_df["Economy..GDP.per.Capita."] >= 1.6, :]
gdp_score.head()

## 9. Select all the Countries Where the Happiness Score is Greater than or Equal to 5 and GDP is Less than or Equal to 1

# Select all rows that have a happiness score greater than or equal 5 and a GDP score less than or equal 1
happy_but_poor_score = happiness_df.loc[(happiness_df["Happiness.Score"] >= 5) & 
                                        (happiness_df["Economy..GDP.per.Capita."] <= 1), :]
happy_but_poor_score.head()
